fix: Move record queueing to a generic method inside of `publish.js`. Use the new queueing functionality in the broadcast endpoint.
