+++ src/index.js
@@ -1,17 +1,9 @@
-/* global AbortController */
 import websocket from 'websocket'
 import dotenv from 'dotenv'
 import debug from 'debug'
-import { create as createIpfs } from 'ipfs-http-client'
-import PQueue from 'p-queue'
-import formatNumber from 'format-number'
-import { publishRecord } from './publish.js'
-import { shorten } from './utils/string.js'
+import { addToQueue } from './publish.js'
 
 dotenv.config()
-
-const CONCURRENCY = 5
-const fmt = formatNumber()
 
 const WebSocket = websocket.client
 const log = debug('ipns-pub')
@@ -27,15 +19,7 @@ async function main () {
   const endpoint = process.env.ENDPOINT || 'wss://api.web3.storage'
   const url = new URL('name/*/watch', endpoint)
 
-  /** @type {Map<string, { record: string }>} */
-  const taskData = new Map()
-  /** @type {Set<string>} */
-  const runningTasks = new Set()
-  const queue = new PQueue({ concurrency: CONCURRENCY })
-
   while (true) {
-    const ipfs = createIpfs()
-
     /** @type {import('websocket').connection} */
     const conn = await new Promise((resolve, reject) => {
       const client = new WebSocket()
@@ -49,47 +33,8 @@ async function main () {
       await new Promise((resolve, reject) => {
         conn.on('message', async msg => {
           const { key, value, record: b64Record } = JSON.parse(msg.utf8Data)
-          const keyLog = log.extend(shorten(key))
-          keyLog.enabled = true
-          keyLog(`🆕 /ipns/${key} ➡️ ${value}`)
 
-          let data = taskData.get(key)
-          if (data) {
-            Object.assign(data, { value, record: b64Record })
-            return keyLog('👌 Already in the queue (record to publish has been updated)')
-          }
-
-          data = { value, record: b64Record }
-          taskData.set(key, data)
-
-          const start = Date.now()
-          keyLog(`➕ Adding to the queue, position: ${fmt(queue.size)}`)
-          queue.add(async function run () {
-            // if this task is already running, lets not concurrently put
-            // multiple versions for the same key!
-            if (runningTasks.has(key)) {
-              keyLog('🏃 Already running! Re-queue in 60s...')
-              await sleep(60_000)
-              if (taskData.has(key) && taskData.get(key) !== data) {
-                return keyLog('⏩ Skipping re-queue, a newer update has been queued already.')
-              }
-              taskData.set(key, data)
-              keyLog(`➕ Re-adding to the queue, position: ${fmt(queue.size)}`)
-              queue.add(run)
-              return
-            }
-            keyLog(`🏁 Starting publish (was queued for ${fmt(Date.now() - start)}ms)`)
-            runningTasks.add(key)
-
-            try {
-              const data = taskData.get(key)
-              if (!data) throw new Error('missing task data')
-              taskData.delete(key)
-              publishRecord(ipfs, key, data.value, data.record)
-            } finally {
-              runningTasks.delete(key)
-            }
-          })
+          await addToQueue(key, value, b64Record)
         })
 
         conn.on('error', err => reject(err))
