+++ src/publish.js
@@ -2,8 +2,15 @@
 import * as uint8arrays from 'uint8arrays'
 import debug from 'debug'
 import formatNumber from 'format-number'
+import PQueue from 'p-queue'
 import { shorten } from './utils/string.js'
+import { create as createIpfs } from 'ipfs-http-client'
 
+const ipfs = createIpfs()
+
+const sleep = ms => new Promise(resolve => setTimeout(resolve, ms))
+
+const CONCURRENCY = 5
 const DHT_PUT_TIMEOUT = 60_000
 const fmt = formatNumber()
 
@@ -11,15 +18,73 @@ const log = debug('ipns-pub')
 log.enabled = true
 log.debug = debug('ipns-pub-debug')
 
+/** @type {Map<string, { record: string }>} */
+const taskData = new Map()
+/** @type {Set<string>} */
+const runningTasks = new Set()
+const queue = new PQueue({ concurrency: CONCURRENCY })
+
 /**
- * Publish the given base64-encoded IPNS record for the given public key to the DHT.
- * @param {object} ipfs IPFS client.
+ * Queues the IPNS record to be added to the DHT.
  * @param {string} key Public key of the record.
  * @param {string} value The `value` field of the record.
  * @param {string} b64record Base 64 encoded serialized IPNS record.
  * @returns {undefined}
  */
-export async function publishRecord(ipfs, key, value, b64record) {
+export async function addToQueue (key, value, b64record) {
+  const keyLog = log.extend(shorten(key))
+  keyLog.enabled = true
+
+  let data = taskData.get(key)
+  if (data) {
+    Object.assign(data, { value, b64record })
+    return keyLog('👌 Already in the queue (record to publish has been updated)')
+  }
+
+  data = { value, b64record }
+  taskData.set(key, data)
+
+  const start = Date.now()
+  keyLog(`➕ Adding to the queue, position: ${fmt(queue.size)}`)
+  queue.add(async function run () {
+    // if this task is already running, lets not concurrently put
+    // multiple versions for the same key!
+    if (runningTasks.has(key)) {
+      keyLog('🏃 Already running! Re-queue in 60s...')
+      await sleep(60_000)
+      if (taskData.has(key) && taskData.get(key) !== data) {
+        return keyLog('⏩ Skipping re-queue, a newer update has been queued already.')
+      }
+
+      // Add back into queue
+      taskData.set(key, data)
+      keyLog(`➕ Re-adding to the queue, position: ${fmt(queue.size)}`)
+      queue.add(run)
+      return
+    }
+
+    keyLog(`🏁 Starting publish (was queued for ${fmt(Date.now() - start)}ms)`)
+    runningTasks.add(key)
+
+    try {
+      const data = taskData.get(key)
+      if (!data) throw new Error('missing task data')
+      taskData.delete(key)
+      publishRecord(key, data.value, data.record)
+    } finally {
+      runningTasks.delete(key)
+    }
+  })
+}
+
+/**
+ * Publish the given base64-encoded IPNS record for the given public key to the DHT.
+ * @param {string} key Public key of the record.
+ * @param {string} value The `value` field of the record.
+ * @param {string} b64record Base 64 encoded serialized IPNS record.
+ * @returns {undefined}
+ */
+export async function publishRecord (key, value, b64record) {
   const keyLog = log.extend(shorten(key))
   const start = Date.now()
   let timeoutId
